feat: partial implementation of hard pitch constraints
We use this when inertia is disabled, but it's glitchy.
At least the camera does not crash anymore :3

This took me way too long for a poor end result.
I do have some more ideas for fixes, stay tuned!

This commit breaks API so we'll go straight to v9
and try to forget the v8 ever existed.

/breaking
/spend 15h